requirements.txt
tensorflow==2.15.0
streamlit==1.35.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
scikit-learn==1.5.0
locust==2.29.0
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
